requirements.txt
# This project uses uv for dependency management
# Install with: pip install uv
# Then run: uv pip install -r requirements.txt

fastapi==0.103.1
uvicorn==0.23.2
pydantic==2.3.0
pydantic-settings==2.0.3
sqlalchemy==2.0.20
alembic==1.12.0
pytest==7.4.2
httpx==0.24.1
python-dotenv==1.0.0
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.0.1
python-multipart==0.0.6
gunicorn==21.2.0
